+++ tests/ICWireTest.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
+
 #include "../src/IC.cpp"
 #include "../src/Wire.cpp"
-#include "../src/ANDGateIC.cpp"
 
 // Concrete IC class for testing that implements simulate()
 template <typename T>
@@ -137,49 +137,6 @@ TEST_F(ICTest, InvalidWireConnection) {
     // Wire to nullptr should throw
     EXPECT_THROW(Wire<int>(&sourceIC, 2, nullptr, 3), runtime_error);
 }
-
-// ----------------------------------------------------------------
-
-// Test fixture for ANDGateIC
-class ANDGateICTest : public ::testing::Test {
-protected:
-    ANDGateIC<bool> andGate;
-};
-
-// Test ANDGateIC Constructor
-TEST_F(ANDGateICTest, Constructor) {
-    EXPECT_EQ(andGate.getTotalPins(), 14);
-    EXPECT_EQ(andGate.getName(), "AND");
-}
-
-// Test Simulate Method
-TEST_F(ANDGateICTest, Simulate) {
-    andGate.connectVCC();
-    andGate.connectGround();
-
-    // Set inputs for Gate 1
-    andGate.setPin(1, true);
-    andGate.setPin(2, true);
-
-    // Simulate and check output
-    EXPECT_NO_THROW(andGate.simulate());
-    EXPECT_TRUE(andGate.getPin(3));
-
-    // Set inputs for Gate 1
-    andGate.setPin(1, true);
-    andGate.setPin(2, false);
-
-    // Simulate and check output
-    andGate.simulate();
-    EXPECT_FALSE(andGate.getPin(3));
-}
-
-// Test Simulate without Power
-TEST_F(ANDGateICTest, SimulateWithoutPower) {
-    EXPECT_THROW(andGate.simulate(), std::runtime_error);
-}
-
-// ----------------------------------------------------------------
 
 int main(int argc, char **argv) {
     ::testing::InitGoogleTest(&argc, argv);
